util::extract_subcontract_manifest_path -> ManifestPath::new_from_subcontract_package_id
