+++ crates/cargo-contract/src/util/mod.rs
@@ -17,10 +17,7 @@
 #[cfg(test)]
 pub mod tests;
 
-use crate::{
-    ManifestPath,
-    Verbosity,
-};
+use crate::Verbosity;
 use anyhow::{
     Context,
     Result,
@@ -168,27 +165,6 @@ pub fn extract_subcontract_name(package_id: PackageId) -> Option<String> {
     let caps = re.captures(package_id.repr.as_str())?;
     let package = caps.get(1)?.as_str();
     Some(String::from(package))
-}
-
-/// PackageId looks like this:
-/// `subcontract 3.0.0 (path+file:///path/to/subcontract)`
-/// so we have to extract the manifest_path via regex:
-pub fn extract_subcontract_manifest_path(package_id: PackageId) -> Result<ManifestPath> {
-    let re = Regex::new(r"\((.*)\)")?;
-    let caps = re.captures(package_id.repr.as_str()).ok_or_else(|| {
-        regex::Error::Syntax("Cannot extract manifest path".to_string())
-    })?;
-    let path_str = caps
-        .get(1)
-        .ok_or_else(|| anyhow::anyhow!("Manifest not extracted"))?
-        .as_str()
-        .replace("path+file://", "");
-
-    let mut path = PathBuf::new();
-    path.push(path_str);
-    path.push("Cargo.toml");
-
-    ManifestPath::try_from(Some(path))
 }
 
 /// Prints to stdout if `verbosity.is_verbose()` is `true`.
