+++ crates/cargo-contract/src/workspace/manifest.rs
@@ -18,16 +18,15 @@ use anyhow::{
     Context,
     Result,
 };
+use cargo_metadata::PackageId;
+use regex::Regex;
 
 use super::{
     metadata,
     Profile,
 };
 use crate::{
-    util::{
-        extract_subcontract_manifest_path,
-        extract_subcontract_name,
-    },
+    util::extract_subcontract_name,
     OptimizationPasses,
 };
 
@@ -87,6 +86,28 @@ impl ManifestPath {
         Ok(manifest_path)
     }
 
+    /// Create a new ['ManifestPath'] from a subcontract PackageId
+    pub fn new_from_subcontract_package_id(package_id: PackageId) -> Result<Self> {
+        // PackageId looks like this:
+        // `subcontract 3.0.0 (path+file:///path/to/subcontract)`
+        // so we have to extract the manifest_path via regex:Result<Self> {
+        let re = Regex::new(r"\((.*)\)")?;
+        let caps = re.captures(package_id.repr.as_str()).ok_or_else(|| {
+            regex::Error::Syntax("Cannot extract manifest path".to_string())
+        })?;
+        let path_str = caps
+            .get(1)
+            .ok_or_else(|| anyhow::anyhow!("Manifest not extracted"))?
+            .as_str()
+            .replace("path+file://", "");
+
+        let mut path = PathBuf::new();
+        path.push(path_str);
+        path.push("Cargo.toml");
+
+        ManifestPath::try_from(Some(path))
+    }
+
     /// Create an arg `--manifest-path=` for `cargo` command
     pub fn cargo_arg(&self) -> Result<String> {
         let path = self.path.canonicalize().map_err(|err| {
@@ -130,7 +151,7 @@ impl ManifestPath {
             .find(|package_id| {
                 extract_subcontract_name(package_id.clone()) == Some(package.to_string())
             })
-            .and_then(|id| extract_subcontract_manifest_path(id).ok())
+            .and_then(|id| ManifestPath::new_from_subcontract_package_id(id).ok())
             .ok_or_else(|| anyhow::anyhow!("subcontract manifest not extracted"))?;
         Ok(manifest_path)
     }
